Another rearrangement, to share provisioning scripts

+++ deploy/provision.d/100-apache-config.sh
@@ -4,13 +4,20 @@ set -e
 
 cd /var/www/code
 
+codeconffile=/var/www/code/deploy/config/code.conf
+
+if [ ! -f "$codeconffile" ]; then
+    echo "ERROR: Apache config file $codeconffile not found - has the database secret been interpolated from $codeconffile.in correctly?"
+    exit 2
+fi
+
 if [ ! -f /etc/apache2/sites-enabled/10-code.conf ]; then
     
     rm -f /etc/apache2/sites-enabled/000-default.conf
 
-    cp extra/soundsoftware/scripted-deploy/config/passenger.conf /etc/apache2/mods-available/
-    cp extra/soundsoftware/scripted-deploy/config/passenger.load /etc/apache2/mods-available/
-    cp extra/soundsoftware/scripted-deploy/config/perl.conf      /etc/apache2/mods-available/
+    cp deploy/config/passenger.conf /etc/apache2/mods-available/
+    cp deploy/config/passenger.load /etc/apache2/mods-available/
+    cp deploy/config/perl.conf      /etc/apache2/mods-available/
 
     ln -s ../mods-available/passenger.conf  /etc/apache2/mods-enabled/
     ln -s ../mods-available/passenger.load  /etc/apache2/mods-enabled/
@@ -19,7 +26,7 @@ if [ ! -f /etc/apache2/sites-enabled/10-code.conf ]; then
     ln -s ../mods-available/rewrite.load    /etc/apache2/mods-enabled/
     ln -s ../mods-available/cgi.load        /etc/apache2/mods-enabled/
 
-    cp extra/soundsoftware/scripted-deploy/config/code.conf.interpolated /etc/apache2/sites-available/code.conf
+    cp "$codeconffile" /etc/apache2/sites-available/code.conf
     ln -s ../sites-available/code.conf /etc/apache2/sites-enabled/10-code.conf
 
     apache2ctl configtest
